moved snippet related functions to different files
